wip: move TaxaList into TaxonSearch
