+++ src/sharedHooks/useKeyboardInfo.js
@@ -5,7 +5,7 @@ import {
 } from "react-native";
 
 // Returns info about the keyboard, like whether it's up and how tall it is
-const useKeyboardInfo = ( targetNonKeyboardHeight: number ) => {
+const useKeyboardInfo = ( targetNonKeyboardHeight = 0 ) => {
   const [keyboardShown, setKeyboardShown] = useState( false );
   const [keyboardHeight, setKeyboardHeight] = useState( 0 );
   const nonKeyboardHeight = useMemo(
@@ -38,6 +38,7 @@ const useKeyboardInfo = ( targetNonKeyboardHeight: number ) => {
   }, [keyboardShown] );
 
   return {
+    keyboardHeight,
     keyboardShown,
     keyboardVerticalOffset,
     nonKeyboardHeight
